Output frametime

+++ src/scene/object.hpp
@@ -31,5 +31,5 @@ class Object {
     NLOHMANN_DEFINE_TYPE_INTRUSIVE(Object, name, meshIdx, shaderIdx, material, rotate, translation, rotation, scale);
 
     unsigned int id;
-    Series<float, Config::NUMBER_OF_MEASUREMENTS> measurements;
+    Series<GLuint, Config::NUMBER_OF_MEASUREMENTS> measurements;
 };
